LFG.gotmpl V2
While I updated this recently, I noticed there were some slight errors with it, and improvements, I felt could be made.

As such, I decided I'd remove the override from the public version. I felt it didn't suit the design of the command, and was a level of needless complication, I put in for my own uses, not for others. I also shuffled round some of the variables, to an order I thought was more suitable, and moved the argument handling, down to be with the rest of the command script. 

Along with this, I also updated the elements of the script, removing the override check, as well as updating the command branches, to fit with the move of the outcome messages.

On that note, the most significant addition to this version of the script, is the definition of the outcome messages, using variables! Previously, they were embedded in the functional portion of the script. Instead, I moved them up into the configuration section, to highlight to users, that these messages were readily customisable.

Finally, I removed some unnecessary comments, I didn't believe fit in the script, and could be a bit more professional.

That concludes an overview of my thoughts and changes, to Version 2 of my LFG script!