Relocate script file and consolidate variable case
As my ethos stands, I generally like to create a new branch, when it is my intention to make a significant change to my repository, and in this case, that is the introduction of an entirely new folder section in the repository, in this case, being a section for moderation commands. To this end, the primary change of this commit, is changing the folder the file will be found under.

Secondly, it also consolidates the case the variables in the script are written in, as prior to this, the variables listed in the configuration section, from when the script was initially refactored for publication, were changed to be written in camel case, which is not my preferred case style, and was the only section in the command, where it was used by choice.
Additional, in subsequent additions, when the $require_manage variable was added, I reverted to using snake case. As such, I took the decision to consolidate the case used in the script, back to my preferred snake case, for all future implementations.

Otherwise, there have corrections to some small been some small grammatical and spelling errors.