Create Custom Role Command Controls.gotmpl
This script, was designed as an significant enhancement to the -give and -takerole commands. This includes:

+ A setting to exclude bot targets.
+ A list of usable arguments for assigning roles.
+ The ability to use role mentions, as well as role names or IDs.
+ Validates duration arguments are within range.

While the bulk of the script is dedicated to the necessary process of validating the executing the command, I'd hope the changes it provides are significant.

By default, the manage messages permission check, simply acts as a default check, on whether a user will be able to use the command, based on their permissions alone, it also plays into two additional configuration settings, buried in the script, and disabled by default. These settings will override the role control, and bot exclusion check if they are turned on, and the user has permission to manage roles, essentially reverting it to the base state, but additionally allowing for the use of role mentions, and handling exceptional durations.
These are disabled by default, to prevent confusion around the command's behaviour, but allowing their location, and use, by users who understand the functionality of the script, at a more advanced level.

Overall, I'm very happy with my work on this script, regardless of the number of people who will actively employ it.