Create Flag Handling Manual Arg Parser.gotmpl
This script has been created as an example, of a basic, but functional, manual argument parsing script, that extracts any switches or flags used in a commands arguments. These are stored in an sdict, with a Boolean or string value, corresponding to a switch or flag argument, respectively.
It's designed not to perform any further operations while extracting values, in order to be ubiquitous and adaptable, allowing for the values to be converted afterwards, instead of trying to account for any type conversions desired, when extracting the values.

As the first script snippet I created, the scripts in this section of my repository, won't serve any immediate and applicable function. Instead, they will serve as example scripts and systems, designed to be ubiquitous.
To this end, they will be designed so they can easily be transplanted into, be adjusted for use or work in conjunction with, other scripts, to quickly enable more comprehensive, and frequently employed functionality, such as enabling the simplified extraction of switch and flag arguments. Further additions to this section, should follow a similar pattern, whether they're a system, or an individual script.