Update Custom Role Command Controls.gotmpl
Still tinkering with this script, because DZ found out a better solution for extracting a RegEx safe variant of the bot prefix, without compiling the entire thing as a string of lowercase hex characters, using reReplace and the quote meta sequence, \Q and \E, to escape any meta behaviour. Gonna say people can still work out the trigger themselves... I don't want to bloat that with confusing and unnecessary characters.