Switch licence to GNU GPL v2.0
Due to the nature of the MIT licence allowing for the modification for, and redistribution of proprietary and monetizable works, which is not my intention. As such, I've taken the decision to switch the repository licence over to GNU GPL v2.0, which better represents the desire to distribute beneficial scripts for YAGPDB, free of charge.